添加调研模块

+++ JJEhr/event/views.py
@@ -1,9 +1,7 @@
 #-*- coding: UTF-8 -*-
-from StringIO import StringIO
 from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse
 from django.shortcuts import render_to_response
-from django.views.decorators.csrf import csrf_protect
 from django.views.decorators.http import require_http_methods
 from JJEhr.event.form import AddEventTypeForm
 from JJEhr.event.models import EventType
@@ -23,31 +21,30 @@ def event_add(request):
     return HttpResponse(result)
 
 
+@require_http_methods(["GET"])
 @login_required(login_url='/backoffice/login')
-@require_http_methods(["POST"])
-@csrf_protect
-def event_delete(request):
-    event_type = EventType.objects.get(id=request.POST['event_type_id'])
-    event_type.delete()
+def event_delete(request, event_type_id=0):
+    if(event_type_id == 0):
+        return  HttpResponse("传入参数异常")
+    try:
+        event_type = EventType.objects.get(id=event_type_id)
+        event_type.delete()
+        return  HttpResponse("删除事件类型成功")
+    except Exception:
+        return  HttpResponse("服务器异常，请稍后再试")
 
 
 @require_http_methods(["GET"])
 @login_required(login_url='/backoffice/login')
-def event_show_all(request):
+def get_all_event_type(request):
     event_type_list = EventType.objects.all()
-    type_name_list = ["<div>" + event_type.type_name + "</div>" for event_type in event_type_list]
-    string_io = StringIO()
-    for type_name in type_name_list:
-        string_io.write(type_name)
-    result = string_io.getvalue()
-    string_io.close()
-    return HttpResponse(result)
+    return render_to_response("ajax/ajaxEditEventContent.html", {"event_type_list": event_type_list})
 
 
 @require_http_methods(["GET"])
 def ajax_content_html(request):
     form = AddEventTypeForm()
-    return render_to_response("ajax/ajax_content.html", {"eventTypeForm": form})
+    return render_to_response("ajax/ajaxAddEventContent.html", {"eventTypeForm": form})
